split xbr db to xbr/xbrmm
