+++ cfxdb/xbrmm/__init__.py
@@ -1,0 +1,62 @@
+##############################################################################
+#
+#                        Crossbar.io FX
+#     Copyright (C) Crossbar.io Technologies GmbH. All rights reserved.
+#
+##############################################################################
+
+from cfxdb.xbrmm.schema import Schema
+
+from cfxdb.xbrmm.consent import Consent, Consents, IndexConsentByMemberAddress
+
+from cfxdb.xbrmm.channel import Channel
+from cfxdb.xbrmm.channel import Channel as PaymentChannel
+from cfxdb.xbrmm.channel import Channel as PayingChannel
+from cfxdb.xbrmm.channel import ChannelBalance
+from cfxdb.xbrmm.channel import ChannelBalance as PaymentChannelBalance
+from cfxdb.xbrmm.channel import ChannelBalance as PayingChannelBalance
+from cfxdb.xbrmm.channel import PaymentChannels, IndexPaymentChannelByDelegate, \
+    PaymentChannelBalances, PayingChannels, IndexPayingChannelByDelegate, \
+    IndexPayingChannelByRecipient, PayingChannelBalances
+
+from cfxdb.xbrmm.offer import Offer, Offers, IndexOfferByKey
+
+from cfxdb.xbrmm.transaction import Transaction, Transactions
+
+from cfxdb.gen.xbr.TransactionState import TransactionState
+from cfxdb.gen.xbr.ChannelType import ChannelType
+from cfxdb.gen.xbr.ChannelState import ChannelState
+
+
+__all__ = (
+    # database schema
+    'Schema',
+
+    # enum types
+    'ChannelType',
+    'ChannelState',
+    'TransactionState',
+
+    # table/index types
+    'Consent',
+    'Consents',
+    'IndexConsentByMemberAddress',
+    'Channel',
+    'PaymentChannel',
+    'PaymentChannels',
+    'IndexPaymentChannelByDelegate',
+    'ChannelBalance',
+    'PaymentChannelBalance',
+    'PaymentChannelBalances',
+    'PayingChannel',
+    'PayingChannels',
+    'IndexPayingChannelByDelegate',
+    'IndexPayingChannelByRecipient',
+    'PayingChannelBalance',
+    'PayingChannelBalances',
+    'Offer',
+    'Offers',
+    'IndexOfferByKey',
+    'Transaction',
+    'Transactions',
+)
