+++ cfxdb/xbrmm/channel.py
